Module Expect_test_collector.Make

Parameters

Config : Expect_test_config.S

Signature

val save_output : Expect_test_common.Std.File.Location.t -> unit Config.IO.t

Collect the output that has been run since the last call to save_output, or since the current expect-test started running.

This function should only be called while a test is running. It is meant to be called as a result of ppx_expect translating an expect-test, and is not intended to be called manually.

val save_and_return_output : Expect_test_common.Std.File.Location.t -> string Config.IO.t

val run : file_digest:Expect_test_common.Std.File.Digest.t -> location:Expect_test_common.Std.File.Location.t -> absolute_filename:string -> description:string option -> tags:string list -> expectations:Expect_test_common.Std.Expectation.Raw.t list -> uncaught_exn_expectation:Expect_test_common.Std.Expectation.Raw.t option -> inline_test_config:Ppx_inline_test_lib.Runtime.config -> (unit -> unit Config.IO.t) -> unit

Run an expect-test